Move ReplaceAnyTypeWithJSON into subpackage
